Feat: 자동로그인 API 구현 / Update: userIdCheck 함수 위치 변경
