+++ src/app/Planner/plannerService.js
@@ -1,10 +1,11 @@
 import baseResponse from '../../../config/baseResponseStatus';
 import pool from '../../../config/database';
 import { response, errResponse } from '../../../config/response';
+import { userIdCheck } from '../User/userProvider';
 import {
   plannerIdCheck,
   scrapIdCheck,
-  userIdCheck,
+  // userIdCheck,
   retrievePlannerId,
   retrieveTimetableId,
 } from './plannerProvider';
