+++ src/app/User/userRoute.js
@@ -1,9 +1,11 @@
 import express from 'express';
-import { loginUser, postUser } from './userController';
+import { jwtMiddleware } from '../../../config/jwtMiddleware';
+import { loginUser, postUser, loginTest } from './userController';
 
 const userRouter = express.Router();
 
 userRouter.post('/login', loginUser);
 userRouter.post('/signup', postUser);
+userRouter.get('/login', jwtMiddleware, loginTest);
 
 export default userRouter;
